・Main画面の測定タップでBmi Calculationから結果を受け取り、その値を結果表示用のActivityに渡して画面遷移する機能の実装 ・結果表示画面でbmiを表示し、タイプに応じて説明文を表示する機能の実装。

+++ BMIApp/app/src/main/java/model/BmiInfo.kt
@@ -1,3 +1,3 @@
 package model
 
-data class BmiInfo (var bmi: Int, var bodyType: String)
+data class BmiInfo (var bmi: Double, var bodyType: String)
